+++ frontend/File.ts
@@ -1,0 +1,15 @@
+
+export interface File {
+    path: string;
+    text: string;
+  }
+
+export function updateFile(file: File, path: string, text: string) {
+    file.path = path;
+    file.text = text;
+}
+
+export function resetFile(resetFile: File) {
+    resetFile.path = "";
+    resetFile.text = "";
+}
